requirements.txt
streamlit
joblib
scikit-learn
pandas
numpy
streamlit_extras.st_modal
